api/requirements.txt
Django >=4.0.1,<4.1
djangorestframework >=3.13.1, <3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
requests
celery
redis